# week-2-assignment-1/swagger.yaml
openapi: 3.0.4
info:
  title: Task Management API
  version: 1.0.0
  description: A simple API for managing tasks

servers:
- url: http://localhost:3000
  description: Local server

tags:
- name: Tasks
description: Operations related to tasks

paths:
    /stats:
        get:
            summary: Get tasks statistics
            tags:
                - Tasks
            description: Retrieve counts of total, completed, and open tasks
            responses:
                "200":
                    description: Tasks statistics
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    total:
                                        type: integer
                                        example: 3
                                    done:
                                        type: integer
                                        example: 1
                                    open:
                                        type: integer
                                        example: 2
    /reset:
        post:
            summary: Reset tasks
            tags:
                - Tasks
            description: Reset tasks to their initial state
            responses:
                "200":
                    description: Tasks reset successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "Tasks reset to initial state"
    /tasks:
        get:
            summary: Get all tasks
            tags: 
                - Tasks
            description: Retrieve a list of all tasks
            parameters:
                - in: query
                  name: done
                  schema:
                      type: string
                  description: Filter tasks by completion status (true/false)
                - in: query
                  name: search
                  schema:
                      type: string
                  description: Filter tasks by title search string
            responses:
                "200":
                    description: A list of tasks
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        id:
                                            type: integer
                                            example: 1
                                        title:
                                            type: string
                                            example: "Sample Task"
                                        done:
                                            type: boolean
                                            example: false
        post:
            summary: Create a new task
            tags: 
                - Tasks
            description: Add a new task to the list
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                    example: "New Task"
            responses:
                "201":
                    description: Task created successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                        example: 2
                                    title:
                                        type: string
                                        example: "New Task"
                                    done:
                                        type: boolean
                                        example: false
                "400":
                    description: Invalid input, task not created
    /tasks/{id}:
        get:
            summary: Get a task by ID
            tags: 
                - Tasks
            description: Retrieve a specific task by its ID
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: integer
                  description: The ID of the task to retrieve
            responses:
                "200":
                    description: Task found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                        example: 1
                                    title:
                                        type: string
                                        example: "Sample Task"
                                    done:
                                        type: boolean
                                        example: false
                "404":
                    description: Task not found
        put:
            description: Update a task by ID
            summary: Update a specific task by its ID
            tags:
                - Tasks
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: integer
                  description: The ID of the task to update
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                    example: "Updated Task"
                                done:
                                    type: boolean
                                    example: true
            responses:
                "200":
                    description: Task updated successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                        example: 1
                                    title:
                                        type: string
                                        example: "Updated Task"
                                    done:
                                        type: boolean
                                        example: true
                "404":
                    description: Task not found
        delete:
            summary: Delete a task by ID
            description: Remove a specific task by its ID
            tags:
                - Tasks
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: integer
                  description: The ID of the task to delete
            responses:
                "204":
                    description: Task deleted successfully
                "404":
                    description: Task not found